chore(math): widen NAT subtractions to Int64

+++ CryptoLib/src/Math/Raw/ClpNat128.pas
@@ -975,16 +975,16 @@ var
   LC: Int64;
 begin
   LC := 0;
-  LC := LC + Int64(AX[0]) - AY[0];
+  LC := LC + (Int64(AX[0]) - Int64(AY[0]));
   AZ[0] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AX[1]) - AY[1];
+  LC := LC + (Int64(AX[1]) - Int64(AY[1]));
   AZ[1] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AX[2]) - AY[2];
+  LC := LC + (Int64(AX[2]) - Int64(AY[2]));
   AZ[2] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AX[3]) - AY[3];
+  LC := LC + (Int64(AX[3]) - Int64(AY[3]));
   AZ[3] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
   Result := Int32(LC);
@@ -996,16 +996,16 @@ var
   LC: Int64;
 begin
   LC := 0;
-  LC := LC + Int64(AX[AXOff + 0]) - AY[AYOff + 0];
+  LC := LC + (Int64(AX[AXOff + 0]) - Int64(AY[AYOff + 0]));
   AZ[AZOff + 0] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AX[AXOff + 1]) - AY[AYOff + 1];
+  LC := LC + (Int64(AX[AXOff + 1]) - Int64(AY[AYOff + 1]));
   AZ[AZOff + 1] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AX[AXOff + 2]) - AY[AYOff + 2];
+  LC := LC + (Int64(AX[AXOff + 2]) - Int64(AY[AYOff + 2]));
   AZ[AZOff + 2] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AX[AXOff + 3]) - AY[AYOff + 3];
+  LC := LC + (Int64(AX[AXOff + 3]) - Int64(AY[AYOff + 3]));
   AZ[AZOff + 3] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
   Result := Int32(LC);
@@ -1016,16 +1016,16 @@ var
   LC: Int64;
 begin
   LC := 0;
-  LC := LC + Int64(AZ[0]) - AX[0] - AY[0];
+  LC := LC + (Int64(AZ[0]) - Int64(AX[0]) - Int64(AY[0]));
   AZ[0] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AZ[1]) - AX[1] - AY[1];
+  LC := LC + (Int64(AZ[1]) - Int64(AX[1]) - Int64(AY[1]));
   AZ[1] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AZ[2]) - AX[2] - AY[2];
+  LC := LC + (Int64(AZ[2]) - Int64(AX[2]) - Int64(AY[2]));
   AZ[2] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AZ[3]) - AX[3] - AY[3];
+  LC := LC + (Int64(AZ[3]) - Int64(AX[3]) - Int64(AY[3]));
   AZ[3] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
   Result := Int32(LC);
@@ -1036,16 +1036,16 @@ var
   LC: Int64;
 begin
   LC := 0;
-  LC := LC + Int64(AZ[0]) - AX[0];
+  LC := LC + (Int64(AZ[0]) - Int64(AX[0]));
   AZ[0] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AZ[1]) - AX[1];
+  LC := LC + (Int64(AZ[1]) - Int64(AX[1]));
   AZ[1] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AZ[2]) - AX[2];
+  LC := LC + (Int64(AZ[2]) - Int64(AX[2]));
   AZ[2] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AZ[3]) - AX[3];
+  LC := LC + (Int64(AZ[3]) - Int64(AX[3]));
   AZ[3] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
   Result := Int32(LC);
@@ -1056,16 +1056,16 @@ var
   LC: Int64;
 begin
   LC := 0;
-  LC := LC + Int64(AZ[AZOff + 0]) - AX[AXOff + 0];
+  LC := LC + (Int64(AZ[AZOff + 0]) - Int64(AX[AXOff + 0]));
   AZ[AZOff + 0] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AZ[AZOff + 1]) - AX[AXOff + 1];
+  LC := LC + (Int64(AZ[AZOff + 1]) - Int64(AX[AXOff + 1]));
   AZ[AZOff + 1] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AZ[AZOff + 2]) - AX[AXOff + 2];
+  LC := LC + (Int64(AZ[AZOff + 2]) - Int64(AX[AXOff + 2]));
   AZ[AZOff + 2] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
-  LC := LC + Int64(AZ[AZOff + 3]) - AX[AXOff + 3];
+  LC := LC + (Int64(AZ[AZOff + 3]) - Int64(AX[AXOff + 3]));
   AZ[AZOff + 3] := UInt32(LC);
   LC := TBitOperations.Asr64(LC, 32);
   Result := Int32(LC);
